services/schema-detection/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.26.2
python-json-logger==2.0.7
json5==0.9.14
httpx==0.26.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pdfkit==1.0.0
openai==1.6.1
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
